bench: remove wasmedge benchmakring for now
