+++ .github/workflows/benchmarks.yml
@@ -15,7 +15,7 @@ jobs:
         run: git submodule update --init --recursive
       - uses: actions-rust-lang/setup-rust-toolchain@v1
         with:
-          toolchain: nightly
+          toolchain: stable
           target: wasm32-wasi
       - name: Setup build env
         run: |
@@ -25,7 +25,7 @@ jobs:
       - name: Run Benchmarks
         run: |
           set -o pipefail
-          cargo +nightly bench -p containerd-shim-benchmarks -- --output-format bencher | tee output.txt
+          cargo bench -p containerd-shim-benchmarks -- --output-format bencher | tee output.txt
       # Download previous benchmark result from cache (if exists)
       - name: Cache benchmark data
         uses: actions/cache@v4
